Renamed jadey back to jadex  to force lowercase dir to git. I know. I know
