+++ test/fixtures/jadex/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["jadex", "flow"]
+}
