--- test/fixtures/jadey/options.json
@@ -1,4 +1,0 @@
-{
-  "sourceType": "module",
-  "plugins": ["jadey", "flow"]
-}
